feat(instrumentation-http):! reduce public API surface by removing exports and making protrected methods private
