+++ experimental/packages/opentelemetry-instrumentation-http/src/index.ts
@@ -16,46 +16,12 @@
 
 export { HttpInstrumentation } from './http';
 export {
-  Err,
-  Func,
-  GetFunction,
-  Http,
-  HttpCallback,
-  HttpCallbackOptional,
   HttpCustomAttributeFunction,
   HttpInstrumentationConfig,
-  HttpRequestArgs,
   HttpRequestCustomAttributeFunction,
   HttpResponseCustomAttributeFunction,
-  Https,
   IgnoreIncomingRequestFunction,
-  IgnoreMatcher,
   IgnoreOutgoingRequestFunction,
-  ParsedRequestOptions,
-  RequestFunction,
-  RequestSignature,
   StartIncomingSpanCustomAttributeFunction,
   StartOutgoingSpanCustomAttributeFunction,
 } from './types';
-export {
-  extractHostnameAndPort,
-  getAbsoluteUrl,
-  getIncomingRequestAttributes,
-  getIncomingRequestAttributesOnResponse,
-  getIncomingRequestMetricAttributes,
-  getIncomingRequestMetricAttributesOnResponse,
-  getOutgoingRequestAttributes,
-  getOutgoingRequestAttributesOnResponse,
-  getOutgoingRequestMetricAttributes,
-  getOutgoingRequestMetricAttributesOnResponse,
-  getRequestInfo,
-  headerCapture,
-  isCompressed,
-  isValidOptionsType,
-  parseResponseStatus,
-  satisfiesPattern,
-  setAttributesFromHttpKind,
-  setRequestContentLengthAttribute,
-  setResponseContentLengthAttribute,
-  setSpanWithError,
-} from './utils';
