+++ experimental/packages/opentelemetry-instrumentation-http/src/utils.ts
@@ -76,7 +76,7 @@ import {
   IgnoreMatcher,
   ParsedRequestOptions,
   SemconvStability,
-} from './types';
+} from './internal-types';
 import forwardedParse = require('forwarded-parse');
 
 /**
